# api/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "coupon-server",
        "contact": {},
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://0.0.0.0:7170/api/v1",
            "variables": {}
        }
    ],
    "paths": {
        "/admin/coupon": {
            "get": {
                "tags": [
                    "Coupon management"
                ],
                "summary": "Get coupon list",
                "operationId": "Getcouponlist",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "",
                        "headers": {}
                    }
                },
                "deprecated": false
            },
            "post": {
                "tags": [
                    "Coupon management"
                ],
                "summary": "Create coupon",
                "operationId": "Createcoupon",
                "parameters": [],
                "requestBody": {
                    "description": "",
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "$ref": "#/components/schemas/CreatecouponRequest"
                                    },
                                    {
                                        "example": {
                                            "coupon_code": "FARMAKO100",
                                            "usage_type": "one_time",
                                            "applicable_medicine_ids": [
                                                "MED99"
                                            ],
                                            "applicable_categories": [
                                                "antibiotic"
                                            ],
                                            "min_order_value": 300,
                                            "valid_from": "2025-05-01T00:00:00+05:30",
                                            "valid_to": "2025-06-01T23:59:59+05:30",
                                            "terms_and_conditions": "Valid on first order only.",
                                            "discount_type": "inventory",
                                            "discount_value": 150,
                                            "max_usage_per_user": 1,
                                            "is_active": 1
                                        }
                                    }
                                ]
                            },
                            "example": {
                                "coupon_code": "FARMAKO100",
                                "usage_type": "one_time",
                                "applicable_medicine_ids": [
                                    "MED99"
                                ],
                                "applicable_categories": [
                                    "antibiotic"
                                ],
                                "min_order_value": 300,
                                "valid_from": "2025-05-01T00:00:00+05:30",
                                "valid_to": "2025-06-01T23:59:59+05:30",
                                "terms_and_conditions": "Valid on first order only.",
                                "discount_type": "inventory",
                                "discount_value": 150,
                                "max_usage_per_user": 1,
                                "is_active": 1
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "",
                        "headers": {}
                    }
                },
                "deprecated": false
            }
        },
...
    },
    "components": {
        "schemas": {
            "CreatecouponRequest": {
                "title": "CreatecouponRequest",
                "required": [
                    "coupon_code",
                    "usage_type",
                    "applicable_medicine_ids",
                    "applicable_categories",
                    "min_order_value",
                    "valid_from",
                    "valid_to",
                    "terms_and_conditions",
                    "discount_type",
                    "discount_value",
                    "max_usage_per_user",
                    "is_active"
                ],
                "type": "object",
                "properties": {
                    "coupon_code": {
                        "type": "string"
                    },
                    "usage_type": {
                        "type": "string"
                    },
                    "applicable_medicine_ids": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "description": ""
                    },
                    "applicable_categories": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "description": ""
                    },
                    "min_order_value": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "valid_from": {
                        "type": "string"
                    },
                    "valid_to": {
                        "type": "string"
                    },
                    "terms_and_conditions": {
                        "type": "string"
                    },
                    "discount_type": {
                        "type": "string"
                    },
                    "discount_value": {
                        "type": "number"
                    },
                    "max_usage_per_user": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "is_active": {
                        "type": "integer",
                        "format": "int32"
                    }
                },
                "example": {
                    "coupon_code": "FARMAKO100",
                    "usage_type": "one_time",
                    "applicable_medicine_ids": [
                        "MED99"
                    ],
                    "applicable_categories": [
                        "antibiotic"
                    ],
                    "min_order_value": 300,
                    "valid_from": "2025-05-01T00:00:00+05:30",
                    "valid_to": "2025-06-01T23:59:59+05:30",
                    "terms_and_conditions": "Valid on first order only.",
                    "discount_type": "inventory",
                    "discount_value": 150,
                    "max_usage_per_user": 1,
                    "is_active": 1
                }
            },
...
        }
    },
...
}